Change create task to dialog
